chore: unsafe key access for ResponseKey
