+++ engine/crates/engine-v2/src/response/read/ser.rs
@@ -2,8 +2,8 @@ use serde::ser::{SerializeMap, SerializeSeq};
 
 use crate::{
     response::{
-        path::UnpackedResponseEdge, GraphqlError, InitialResponse, RequestErrorResponse, ResponseData, ResponseKeys,
-        ResponseObject, ResponsePath, ResponseValue,
+        GraphqlError, InitialResponse, RequestErrorResponse, ResponseData, ResponseKeys, ResponseObject, ResponsePath,
+        ResponseValue, UnpackedResponseEdge,
     },
     Response,
 };
@@ -119,7 +119,7 @@ impl<'a> serde::Serialize for SerializableResponsePath<'a> {
                 UnpackedResponseEdge::BoundResponseKey(key) => {
                     seq.serialize_element(&self.keys.try_resolve(key.into()).unwrap_or("<unknown>"))?
                 }
-                UnpackedResponseEdge::ExtraField(key) => {
+                UnpackedResponseEdge::ExtraFieldResponseKey(key) => {
                     seq.serialize_element(&self.keys.try_resolve(key).unwrap_or("<unknown>"))?
                 }
             }
